Move constants package to shared/
