+++ api/src/main/java/com/example/echo_api/modules/post/api/PostManagementController.java
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.echo_api.constants.ApiRoutes;
 import com.example.echo_api.modules.post.dto.request.CreatePostDTO;
 import com.example.echo_api.modules.post.service.PostManagementService;
+import com.example.echo_api.shared.constants.ApiRoutes;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
