+++ api/src/main/java/com/example/echo_api/modules/post/api/PostViewController.java
@@ -7,9 +7,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.echo_api.constants.ApiRoutes;
 import com.example.echo_api.modules.post.dto.response.PostDTO;
 import com.example.echo_api.modules.post.service.PostViewService;
+import com.example.echo_api.shared.constants.ApiRoutes;
 import com.example.echo_api.shared.dto.PageDTO;
 import com.example.echo_api.shared.validation.pagination.annotations.Limit;
 import com.example.echo_api.shared.validation.pagination.annotations.Offset;
