+++ api/src/main/java/com/example/echo_api/modules/profile/api/ProfileManagementController.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.echo_api.constants.ApiRoutes;
 import com.example.echo_api.modules.profile.dto.request.UpdateProfileDTO;
 import com.example.echo_api.modules.profile.service.ProfileManagementService;
+import com.example.echo_api.shared.constants.ApiRoutes;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
