+++ api/src/test/java/com/example/echo_api/modules/clerk/api/ClerkControllerIT.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 
-import com.example.echo_api.constants.ApiRoutes;
 import com.example.echo_api.modules.user.entity.User;
+import com.example.echo_api.shared.constants.ApiRoutes;
 import com.example.echo_api.testing.support.AbstractIntegrationTest;
 import com.example.echo_api.testing.support.ClerkTestUtils.Template;
