+++ api/src/test/java/com/example/echo_api/modules/profile/api/ProfileManagementControllerIT.java
@@ -10,8 +10,8 @@ import org.springframework.http.HttpStatus;
 
 import com.example.echo_api.config.ErrorMessageConfig;
 import com.example.echo_api.config.ValidationMessageConfig;
-import com.example.echo_api.constants.ApiRoutes;
 import com.example.echo_api.modules.profile.dto.request.UpdateProfileDTO;
+import com.example.echo_api.shared.constants.ApiRoutes;
 import com.example.echo_api.shared.dto.ErrorDTO;
 import com.example.echo_api.testing.support.AbstractIntegrationTest;
