+++ api/src/test/java/com/example/echo_api/modules/profile/api/ProfileViewControllerIT.java
@@ -11,11 +11,11 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpStatus;
 
 import com.example.echo_api.config.ErrorMessageConfig;
-import com.example.echo_api.constants.ApiRoutes;
 import com.example.echo_api.modules.profile.dto.response.ProfileDTO;
 import com.example.echo_api.modules.profile.dto.response.SimplifiedProfileDTO;
 import com.example.echo_api.modules.profile.entity.Follow;
 import com.example.echo_api.modules.profile.repository.FollowRepository;
+import com.example.echo_api.shared.constants.ApiRoutes;
 import com.example.echo_api.shared.dto.ErrorDTO;
 import com.example.echo_api.shared.dto.PageDTO;
 import com.example.echo_api.testing.support.AbstractIntegrationTest;
