+++ api/src/test/java/com/example/echo_api/security/ClerkOnboardingFilterTest.java
@@ -28,7 +28,7 @@ import org.springframework.security.oauth2.jwt.Jwt;
 
 import com.example.echo_api.config.ClerkConfig;
 import com.example.echo_api.config.ErrorMessageConfig;
-import com.example.echo_api.constants.ApiRoutes;
+import com.example.echo_api.shared.constants.ApiRoutes;
 
 import jakarta.servlet.FilterChain;
